+++ tests/test_plugins.py
@@ -1,0 +1,5 @@
+from yast import Yast
+
+
+def test_plugin_init():
+    Yast(plugins={"database": {}})