+++ yast/utils.py
@@ -1,2 +1,0 @@
-def get_modules(module_name: str):
-    pass