+++ changelog.md
@@ -1,6 +1,7 @@
 # v0.2.5
 1. Opt: Add ImmutableMultiDict, Auth: Request -> HttpConnection, graphql: executor_class
 2. database support: mysql
+3. applications: auto scan plugins
 
 # v0.2.4
 1. support template: default jinja2